fix the proto feature
